Polylabel

Documentation for Polylabel.

Polylabel.Polylabel

Polylabel.Cell

Polylabel.polylabel

Polylabel.Polylabel — Module

Polylabel.jl

This package implements an algorithm to find the pole of inaccessibility of a polygon, the most distant internal point from the polygon outline.

This algorithm was originally written (and taken from) mapbox/polylabel

you can find a lot more information there!

The package is built on top of the GeoInterface, and is built to work with any polygons or multipolygon object which implements the GeoInterface, specifically the functions GeoInterface.contains, GeoInterface.distance, and GeoInterface.centroid.

The main entry point is the polylabel(input; rtol, atol) function. It returns a 2-Tuple of floats, representing the x and y coordinates of the pole of inaccessibility that was computed.

Polylabel.Cell — Type

Cell(x, y, half_size, polygon)

Comparison operators operate on the max_distance field.

Polylabel.polylabel — Method

polylabel(polygon::Polygon; rtol::Real = 0.01, atol::Union{Nothing, Real} = nothing)
polylabel(multipoly::MultiPolygon; rtol::Real = 0.01, atol::Union{Nothing, Real} = nothing)

rtol is relative tolerance,