Polylabel

Documentation for Polylabel.

Polylabel.Polylabel

Polylabel.Cell

Polylabel._distance

Polylabel.polylabel

Polylabel.signed_distance

Polylabel.Polylabel — Module

Polylabel.jl

This package implements an algorithm to find the pole of inaccessibility of a polygon, the most distant internal point from the polygon outline.

This algorithm was originally written (and taken from) mapbox/polylabel

you can find a lot more information there!

The package is built on top of the GeoInterface, and is built to work with any polygons or multipolygon object which implements the GeoInterface, specifically the functions GeoInterface.contains, GeoInterface.distance, and GeoInterface.centroid.

The main entry point is the polylabel(input; rtol, atol) function. It returns a 2-Tuple of floats, representing the x and y coordinates of the pole of inaccessibility that was computed.

Polylabel.Cell — Type

Cell(x, y, half_size, polygon)

Comparison operators operate on the max_distance field.

Polylabel._distance — Method

Distance from p0 to the line segment formed by p1 and p2. Implementation from Turf.jl.

Polylabel.polylabel — Method

polylabel(polygon::Polygon; rtol::Real = 0.01, atol::Union{Nothing, Real} = nothing)::Tuple{Float64, Float64}
polylabel(multipoly::MultiPolygon; rtol::Real = 0.01, atol::Union{Nothing, Real} = nothing)::Tuple{Float64, Float64}

polylabel finds the pole of inaccessibility of the given polygon or multipolygon, and returns its coordinates as a 2-Tuple of (x, y). Tolerances can be specified.

Any geometry which expresses the GeoInterface.jl polygon or multipolygon traits can be passed to this method, so long as it implements the GeoInterface methods extent, contains, and centroid, in addition to the polygon coordinates, getexterior, and gethole interfaces.

rtol is relative tolerance, atol is absolute tolerance (in the same vein as Base.isapprox). When atol is provided, it overrides rtol.

Warning

The performance of this function is still being actively improved; specifically the signed distance function needs some optimization. Until then, this will be much slower than the equivalent in Python/JS.

Polylabel.signed_distance — Method

signed_distance(geom, x::Real, y::Real)::Float64

Calculates the signed distance from the geometry geom to the point defined by (x, y). Points within geom have a negative distance, and points outside of geom have a positive distance.

If geom is a MultiPolygon, then this function returns the maximum distance to any of the polygons in geom.